Some modules, function and methods renamed.
