+++ fda/math.py
